Ajout du default de toilet paper au bon endroit
